Glass Properties 1st Draft
This is a first draft at implementing a more declarative property
system.

Any object with properties should inherit from `Glass::HasProperties<T>`
and declare a number of member `PropertyDefinition<T,U,V>` subclasses
and a `PropertyList` named `Properties` that lists all properties to be
created.

Currently this looks like a lot of boiler-plate, but when we are happy
with the implementation, we can make macros to declare properties to
reduce the boiler-plate.

At this point serialization/deserialization has not been implemented.