Update LayoutProperty to call SetNeedsLayout on parent
The purpose of LayoutProperty, DisplayProperty, and UIProperty, was to
make it easy to avoid bugs with not correctly updating in didSet
methods.  LayoutProperty was a potential trap because some properties
that would be considered "layout" would require a layout pass on the
parent (e.g. something that effects intrinsic size calculations).

LayoutProperty was updated to also call SetNeedsLayout on the parent if
it exists.