UI-302/UI-304 (GLAS/glass#189 GLAS/glass#190) Burlap::DesignInterface
UI-302 Burlap::DesignInterface Glass implementation

Implemented two Burlap::DesignInterface subclasses,
Glass::ViewDesignInterface and Burlap::View::ViewDesignInterface, to
provide DesignAid and LayoutFile support for Glass::View and
Burlap::View.

Updated Burlap::BackingView to provide a Glass::ViewDesignInterface and
Glass::View to provide getters/setters for layoutIdentifer and debugName
needed for this DesignInterface implementation.

Updated Glass::HasPropertiesBase to allow subclasses to provide a
Util::PropertyHolder; this is so that Burlap::View can use the wrapped
UI::Window's property holder instead of the PropertyHolder that would be
created by HasPropertiesBase.

Updated Burlap::View to provide a Burlap::internal::ViewDesignInterface
as the wrapped UI::View's DesignInterface and to setup layoutIdentifer.

UI-304 Support Views with fixed position

When a Glass::View has a parent using FlexBox for layout, it should not
be possible to set position via the DesignAid or drag the position of
the View when the design aid is active.

Adds an override for isFixedPosition to Glass::ViewDesignInterface; position is fixed
when in a parent using Flex.